+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `DatabaseConfig#connect` now is deprecated. It is recommended to use `DatabaseConfig#database` property
 * Fixed issue with `PostsSchedulesTable`, when table was not created automatically
 * Refactor work with posts and posts messages
+* `SafeLazy` util class was added
 
 ### 1.8.1